Add defaults to yosysSynthesisReport, fix indentation in openriscPipeline
We should try to have a convention that for synthesis the target is
"synth", we can provide a default for that.  Also, since the logPath is
created by fusesoc we can predict the log location so have a default for
that too.

This will make writing pipelines even more concise.  I am not sure how
to test this.